instances: migrate startup modules into package
