+++ src/smi_beamline/instances/__init__.py
@@ -23,21 +23,21 @@ import time
 # the bootstrap modules base/base_dev which the thin startup runs first to create the session).
 # Each entry: (label, module_name).
 DEVICE_MODULES = [
-    ("sample chamber", "smibase.waxschamber"),
+    ("sample chamber", "smi_beamline.instances.waxschamber"),
     ("shutters",       "smibase.shutter"),
-    ("beamstop",       "smibase.beamstop"),
-    ("machine/ring",   "smibase.machine"),
+    ("beamstop",       "smi_beamline.instances.beamstop"),
+    ("machine/ring",   "smi_beamline.instances.machine"),
     ("attenuators",    "smibase.attenuators"),
-    ("CRLs",           "smibase.crls"),
+    ("CRLs",           "smi_beamline.instances.crls"),
     ("manipulators",   "smibase.manipulators"),
     ("mirrors",        "smibase.mirrors"),
-    ("motors",         "smibase.motors"),
-    ("slits",          "smibase.slits"),
+    ("motors",         "smi_beamline.instances.motors"),
+    ("slits",          "smi_beamline.instances.slits"),
     ("energy",         "smibase.energy"),
-    ("xbpms",          "smibase.xbpms"),
-    ("ioLogik",        "smibase.ioLogik"),
-    ("electrometers",  "smibase.electrometers"),
-    ("amptek",         "smibase.amptek"),
+    ("xbpms",          "smi_beamline.instances.xbpms"),
+    ("ioLogik",        "smi_beamline.instances.ioLogik"),
+    ("electrometers",  "smi_beamline.instances.electrometers"),
+    ("amptek",         "smi_beamline.instances.amptek"),
     ("pilatus",        "smibase.pilatus"),
     ("prosilica",      "smibase.prosilica"),
     ("beam modes",     "smibase.beam"),
@@ -45,9 +45,9 @@ DEVICE_MODULES = [
     ("config",         "smi_beamline.plans.config"),
     ("scan naming",    "smi_beamline.plans.scan_naming"),
     ("beam snapshots", "smi_beamline.plans.beam_snapshot"),
-    ("blade coater",   "smibase.bladecoater"),
+    ("blade coater",   "smi_beamline.instances.bladecoater"),
     ("humidity cell",  "smi_beamline.plans.humidity_cell"),
-    ("linkam",         "smibase.linkam"),
+    ("linkam",         "smi_beamline.instances.linkam"),
     ("suspenders",     "smibase.suspenders"),
     ("utils",          "smi_beamline.plans.utils"),
 ]
