+++ startup/smibase/crls.py
@@ -1,11 +1,1 @@
-
-
-from smi_beamline.devices.crls import CRL
-
-crl = CRL("XF:12IDC-OP:2{Lens:CRL-Ax:", name="crl")
-
-
-from smi_beamline.devices import _context
-
-_context.baseline_register([crl])
-
+from smi_beamline.instances.crls import *
