+++ startup/smibase/linkam.py
@@ -1,5 +1,1 @@
-
-from smi_beamline.devices.linkam import Linkam, LinkamTensile, LinkamThermal
-
-LThermal = LinkamThermal("XF:12ID-ES{LINKAM}:", name="LinkamThermal")
-LTensile = LinkamTensile("XF:12ID-ES:{LINKAM}:", name="LinkamTensile")
+from smi_beamline.instances.linkam import *
